# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/companies/{companyId}/members": {
      "get": {
        "description": "Xem danh sách thành viên của một công ty.",
        "operationId": "CompanyMembersController_listMembers",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "path",
            "description": "Company UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Company members fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CompanyMember"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Company not found"
          }
        },
        "summary": "Danh sách thành viên công ty",
        "tags": ["Company - Members"]
      }
    },
    "/companies/{companyId}/members/invite": {
      "post": {
        "description": "Mời recruiter vào công ty bằng email.",
        "operationId": "CompanyMembersController_inviteMember",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "path",
            "description": "Company UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteMemberDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Invitation sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyMemberInvitation"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload"
          },
          "404": {
            "description": "Company or recruiter account not found"
          },
          "409": {
            "description": "Recruiter is already a member"
          }
        },
        "summary": "Mời thành viên vào công ty",
        "tags": ["Company - Members"]
      }
    },
    "/company-members/invitations/{id}/accept": {
      "post": {
        "description": "Recruiter chấp nhận lời mời tham gia công ty.",
        "operationId": "CompanyMembersController_acceptInvitation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Company member (invitation) UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invitation accepted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyMemberInvitationStatus"
                }
              }
            }
          },
          "404": {
            "description": "Invitation not found"
          },
          "409": {
            "description": "Invitation is not in INVITED status"
          }
        },
        "summary": "Chấp nhận lời mời vào công ty",
        "tags": ["Company - Members"]
      }
    },
    "/company-members/{id}/role": {
      "patch": {
        "description": "Đổi vai trò của thành viên trong công ty.",
        "operationId": "CompanyMembersController_updateRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Company member UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Member role updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyMemberRoleUpdate"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload"
          },
          "404": {
            "description": "Member or role not found"
          }
        },
        "summary": "Cập nhật vai trò thành viên",
        "tags": ["Company - Members"]
      }
    },
    "/company-members/{id}": {
      "delete": {
        "description": "Xóa thành viên khỏi công ty.",
        "operationId": "CompanyMembersController_removeMember",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Company member UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Member removed successfully"
          },
          "404": {
            "description": "Member not found"
          }
        },
        "summary": "Xóa thành viên khỏi công ty",
        "tags": ["Company - Members"]
      }
    },
...
    "/employment-types": {
      "post": {
        "operationId": "EmploymentTypesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmploymentTypeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Tạo loại hình công việc",
        "tags": ["Employment - Types"]
      },
      "get": {
        "operationId": "EmploymentTypesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Danh sách loại hình công việc",
        "tags": ["Employment - Types"]
      }
    },
    "/employment-types/{id}": {
      "get": {
        "operationId": "EmploymentTypesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Chi tiết loại hình công việc",
        "tags": ["Employment - Types"]
      },
      "patch": {
        "operationId": "EmploymentTypesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmploymentTypeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Cập nhật loại hình công việc",
        "tags": ["Employment - Types"]
      },
      "delete": {
        "operationId": "EmploymentTypesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Xóa loại hình công việc",
        "tags": ["Employment - Types"]
      }
    },
    "/experience-levels": {
      "post": {
        "operationId": "ExperienceLevelsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExperienceLevelDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Tạo cấp độ kinh nghiệm",
        "tags": ["Experience - Levels"]
      },
      "get": {
        "operationId": "ExperienceLevelsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Danh sách cấp độ kinh nghiệm",
        "tags": ["Experience - Levels"]
      }
    },
    "/experience-levels/{id}": {
      "get": {
        "operationId": "ExperienceLevelsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Chi tiết cấp độ kinh nghiệm",
        "tags": ["Experience - Levels"]
      },
      "patch": {
        "operationId": "ExperienceLevelsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateExperienceLevelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Cập nhật cấp độ kinh nghiệm",
        "tags": ["Experience - Levels"]
      },
      "delete": {
        "operationId": "ExperienceLevelsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Xóa cấp độ kinh nghiệm",
        "tags": ["Experience - Levels"]
      }
    },
...
  },
  "info": {
    "title": "UpNext API",
    "description": "Backend API for the UpNext IT recruitment platform.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CompanyMemberProfile": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "example": "Nguyen Van A"
          },
          "avatarUrl": {
            "type": "object",
            "nullable": true,
            "example": null
          }
        },
        "required": ["fullName"]
      },
      "CompanyMemberRecruiterAccount": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "b11eaeff-087f-4677-b8bd-c29ac7e59693"
          },
          "email": {
            "type": "string",
            "example": "recruiter@company.com"
          },
          "status": {
            "type": "string",
            "example": "ACTIVE"
          },
          "profile": {
            "$ref": "#/components/schemas/CompanyMemberProfile"
          }
        },
        "required": ["id", "email"]
      },
      "CompanyMemberRole": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "f2d32130-4f55-4517-b8f4-f0ac59a8b2cb"
          },
          "code": {
            "type": "string",
            "example": "hr_manager"
          },
          "name": {
            "type": "string",
            "example": "HR Manager"
          }
        },
        "required": ["id", "code", "name"]
      },
      "CompanyMember": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1f5f4a65-50d7-4f24-a65f-4f2a4d42f9cf"
          },
          "status": {
            "type": "string",
            "enum": ["ACTIVE", "INVITED", "REMOVED"],
            "example": "ACTIVE"
          },
          "joinedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-06-09T08:00:00.000Z"
          },
          "invitedEmail": {
            "type": "object",
            "example": "recruiter@company.com",
            "nullable": true
          },
          "recruiterAccount": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/CompanyMemberRecruiterAccount"
              }
            ]
          },
          "role": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/CompanyMemberRole"
              }
            ]
          }
        },
        "required": ["id", "status", "joinedAt"]
      },
      "InviteMemberDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "recruiter@company.com"
          },
          "roleId": {
            "type": "object",
            "nullable": true,
            "example": null,
            "description": "Existing recruiter role UUID. Omit to invite without a role."
          }
        },
        "required": ["email"]
      },
      "CompanyMemberInvitation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1f5f4a65-50d7-4f24-a65f-4f2a4d42f9cf"
          },
          "status": {
            "type": "string",
            "enum": ["ACTIVE", "INVITED", "REMOVED"],
            "example": "INVITED"
          },
          "joinedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-06-09T08:00:00.000Z"
          },
          "invitedEmail": {
            "type": "string",
            "example": "recruiter@company.com"
          },
          "recruiterAccount": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/CompanyMemberRecruiterAccount"
              }
            ]
          },
          "role": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/CompanyMemberRole"
              }
            ]
          }
        },
        "required": ["id", "status", "joinedAt", "invitedEmail"]
      },
      "CompanyMemberInvitationStatus": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1f5f4a65-50d7-4f24-a65f-4f2a4d42f9cf"
          },
          "status": {
            "type": "string",
            "enum": ["ACTIVE", "INVITED", "REMOVED"],
            "example": "ACTIVE"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-06-09T09:00:00.000Z"
          }
        },
        "required": ["id", "status", "updatedAt"]
      },
      "UpdateMemberRoleDto": {
        "type": "object",
        "properties": {
          "roleId": {
            "type": "string",
            "example": "b11eaeff-087f-4677-b8bd-c29ac7e59693"
          }
        },
        "required": ["roleId"]
      },
      "CompanyMemberRoleUpdate": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1f5f4a65-50d7-4f24-a65f-4f2a4d42f9cf"
          },
          "roleId": {
            "type": "string",
            "example": "f2d32130-4f55-4517-b8f4-f0ac59a8b2cb"
          },
          "role": {
            "$ref": "#/components/schemas/CompanyMemberRole"
          }
        },
        "required": ["id", "roleId", "role"]
      },
...
      "CreateEmploymentTypeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": ["name"]
      },
      "UpdateEmploymentTypeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "CreateExperienceLevelDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "description": "Unique code e.g. JUNIOR, SENIOR"
          }
        },
        "required": ["name", "code"]
      },
      "UpdateExperienceLevelDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "description": "Unique code e.g. JUNIOR, SENIOR"
          }
        }
      },
...
    }
  }
}